Added ban.py cog
Removed ban commands from main.py and moved them to ban.py cog
updated .gitignore
updated counter.py reset_tally description
updated custom1.py cooldowns
added try.py
