+++ cogs/try.py
@@ -1,0 +1,6 @@
+"""
+This file should have nothing in it.
+It is opened during startup on load_cogs.
+If this file fails to load during startup,
+-- that means our load and unload functions should be working correctly.
+"""
